Ugly, but we scope the list items correctly into a component
this is a strange inversion, since the controller for a list is the list view it is contained
in, but the list view is a view component of another UiComponent which most likely
has a fragment controller.

As a result, we have this awkward Dagger within Daggers stuff going on.