Don't use insert conflict strategy REPLACE
REPLACE does not tell us accurately if a new row was added or updated
because room internally uses auto row ids.

We instead have to manually perform the query-insert-update flow
ourselves via the DAO, and we can return a boolean to convey Inserted vs
Updated, or Deletion success, or stuff like that.

A boolean is still ambiguous, we may want an enum in the future for
Inserted vs Updated